requirements.txt
-i https://pypi.org/simple
aiofiles==22.1.0
anyio==3.6.1 ; python_full_version >= '3.6.2'
appdirs==1.4.4
appnope==0.1.3 ; sys_platform == 'darwin'
argcomplete==2.0.0 ; python_version >= '3.6'
argh==0.26.2
asttokens==2.0.8
asyncpg==0.26.0
attrs==22.1.0 ; python_version >= '3.5'
backcall==0.2.0
beautifulsoup4==4.11.1 ; python_full_version >= '3.6.0'
biocommons.seqrepo==0.6.5
bioutils==0.5.7 ; python_version >= '3.6'
boto3==1.24.90
botocore==1.27.90 ; python_version >= '3.7'
bs4==0.0.1
canonicaljson==1.6.3 ; python_version >= '3.7'
certifi==2022.9.24 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
coloredlogs==15.0.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
configparser==5.3.0 ; python_version >= '3.7'
cssselect==1.1.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
decorator==5.1.1 ; python_version >= '3.5'
executing==1.1.1
fake-useragent==0.1.11
fastapi==0.85.0
ga4gh.vrs[extras]==0.8.7.dev0 ; python_version >= '3.6'
ga4gh.vrsatile.pydantic==0.1.dev3 ; python_version >= '3.8'
gene-normalizer==0.2.3
gffutils==0.11.1
h11==0.14.0 ; python_version >= '3.7'
hgvs==1.5.2
humanfriendly==10.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==5.0.0 ; python_version >= '3.7'
inflection==0.5.1 ; python_version >= '3.5'
ipython==8.5.0 ; python_version >= '3.8'
jedi==0.18.1 ; python_version >= '3.6'
jmespath==1.0.1 ; python_version >= '3.7'
jsonschema==3.2.0
lxml==4.9.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markdown==3.4.1 ; python_version >= '3.7'
matplotlib-inline==0.1.6 ; python_version >= '3.5'
numpy==1.23.4 ; python_version >= '3.10'
pandas==1.5.0
parse==1.19.0
parsley==1.3
parso==0.8.3 ; python_version >= '3.6'
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.31 ; python_full_version >= '3.6.2'
psycopg2==2.9.4 ; python_version >= '3.6'
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.2
pyee==8.2.2
pyfaidx==0.7.1
pygments==2.13.0 ; python_version >= '3.6'
pyliftover==0.4
pyppeteer==1.0.2 ; python_version >= '3.7' and python_version < '4.0'
pyquery==1.4.3
pyrsistent==0.18.1 ; python_version >= '3.7'
pysam==0.19.1
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-jsonschema-objects==0.4.1
pytz==2022.4
pyyaml==6.0 ; python_version >= '3.6'
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
requests-html==0.10.0 ; python_full_version >= '3.6.0'
s3transfer==0.6.0 ; python_version >= '3.7'
setuptools==65.5.0 ; python_version >= '3.7'
simplejson==3.17.6 ; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
soupsieve==2.3.2.post1 ; python_version >= '3.6'
sqlparse==0.4.3 ; python_version >= '3.5'
stack-data==0.5.1
starlette==0.20.4 ; python_version >= '3.7'
tabulate==0.9.0 ; python_version >= '3.7'
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
traitlets==5.4.0 ; python_version >= '3.7'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.12 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
uvicorn==0.18.3
w3lib==2.0.1 ; python_version >= '3.6'
wcwidth==0.2.5
websockets==10.3 ; python_version >= '3.7'
yoyo-migrations==8.0.0
zipp==3.9.0 ; python_version >= '3.7'
